+++ lib/twitter/error/already_posted.rb
@@ -1,0 +1,10 @@
+require 'twitter/error/forbidden'
+
+module Twitter
+  class Error
+    # Raised when a Tweet has already been posted
+    class AlreadyPosted < Twitter::Error
+      MESSAGE = "Status is a duplicate"
+    end
+  end
+end